+++ pkg/R/get_raw_api_content.R
@@ -16,7 +16,7 @@
 #' @export
 get_raw_api_content <- function(symbol,
                            api_fun,
-                           api_key = "4HQVKYYYNZTDVUJ2",
+                           api_key = "8ZQHIMUKAYE89GEH",
                            data_type = "json",
                            ...) {