+++ pkg/NEWS.md
@@ -1,6 +1,7 @@
 # fscorescreen 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.
-* Added `get_raw_api_content` that returns raw api content from alphavantage
-* Added `get_f_score_statistics` that calculates the key statistics for f-score screening
-* Added `do_f_score_screening` that performes f-score screening on given ticker, returns `TRUE` if the screening passed.
+* Added `get_raw_api_content()` that returns raw api content from alphavantage
+* Added `get_f_score_statistics()` that calculates the key statistics for f-score screening
+* Added `do_f_score_screening()` that performes f-score screening on given ticker, returns `TRUE` if the screening passed
+* Added `find_target_stocks()` to do f-score screening iteratively over a list of tickers